requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
markdown>=3.4.0
python-dateutil>=2.8.0